\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{
Automates logistic and Mantel-Haenszel DIF detection
}
\description{
Automates logistic and Mantel-Haenszel DIF detection
}
\details{
\tabular{ll}{
Package: \tab DIFdetect\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-07-30\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Tyler Rinker <tyler.rinker@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
